Revert "Merge branch 'main' of https://github.com/sopra-fs23-group-32/SoPra23_Server"
This reverts commit 85f74a37ff4a86a87b6e99771dd4ab3e8154a3c4, reversing
changes made to 82fac79c4e2d31b9abccead9ce62858506adb317.

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/service/GameService.java
@@ -197,10 +197,10 @@ public class GameService {
 
     public void closeGame(Long gameId) {
         Game game = searchGameById(gameId);
-//        if(!game.isGameEnded()) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-//                String.format("Game with ID %d has not finished yet!\n", gameId));
-//        }
+        if(!game.isGameEnded()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                String.format("Game with ID %d has not finished yet!\n", gameId));
+        }
         game.setGameStatus(GameStatus.ENDED);
         updateGameStatus(gameId, WebSocketType.GAME_END, game.getGameStatus());
         gameRepository.delete(game);
